requirements.txt
selenium==4.34.2
webdriver-manager==4.0.2
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4==4.13.4
openai==1.97.1
schedule==1.2.2 